requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.39.0
firebase-admin==6.6.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.31.0